requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
cachelib==0.12.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.3.2
chatbot==1.5.2b0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==42.0.5
cssselect2==0.7.0
cupshelpers==1.0
dataclasses-json==0.6.5
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.4
django-crispy-forms==2.1
djangorestframework==3.15.1
dnspython==2.6.1
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
eval-type-backport==0.1.3
exceptiongroup==1.2.1
fastapi==0.110.3
fasteners==0.14.1
filelock==3.13.3
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.2
greenlet==3.0.3
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
keyring==23.5.0
langchain-community==0.0.34
langchain-core==0.1.46
langsmith==0.1.52
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.1
macaroonbakery==1.3.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.4
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==0.27.8
orjson==3.10.1
oscrypto==1.3.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==16.0.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyGObject==3.42.1
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==4.2.0
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.21.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.4.16
reportlab==4.0.9
requests==2.31.0
rich==13.7.1
routing==0.2.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
sqlparse==0.5.0
starlette==0.37.2
svglib==1.5.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
texttable==1.7.0
tinycss2==1.2.1
together==1.1.2
tokenizers==0.19.1
tqdm==4.66.2
transformers==4.40.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.2
urllib3==2.0.4
usb-creator==0.3.7
uvicorn==0.29.0
virtualenv==20.25.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==3.0.1
xdg==5
xhtml2pdf==0.2.15
xkit==0.0.0
yarl==1.9.4
zipp==3.18.1
